Refactor kind checking logic out of the type checking file.
