+++ gcl.cabal
@@ -27,6 +27,7 @@ library
       GCL.Exec.ExecMonad
       GCL.Exec.ExNondet
       GCL.Exec.ExRand
+      GCL.Kind
       GCL.Predicate
       GCL.Predicate.Located
       GCL.Predicate.Util
@@ -193,6 +194,7 @@ test-suite gcl-test
       GCL.Exec.ExecMonad
       GCL.Exec.ExNondet
       GCL.Exec.ExRand
+      GCL.Kind
       GCL.Predicate
       GCL.Predicate.Located
       GCL.Predicate.Util
