+++ src/Render/Error.hs
@@ -6,7 +6,7 @@ import           Data.Foldable                  ( toList )
 import           Data.Loc                       ( locOf )
 import           Data.Loc.Range
 import           Error
-import           GCL.Type                       ( TypeError(..) )
+import           GCL.Common                     ( TypeError(..) )
 import           GCL.WP.Type                    ( StructError(..) )
 import           Render.Class
 import           Render.Element
