+++ src/Server/Handler/Diagnostic.hs
@@ -13,7 +13,7 @@ import           Error                          ( Error(..) )
 import           GCL.Predicate                  ( Origin(..)
                                                 , PO(..)
                                                 )
-import           GCL.Type                       ( TypeError(..) )
+import           GCL.Common                     ( TypeError(..) )
 import           GCL.WP.Type                    ( StructError(..)
                                                 , StructWarning(..)
                                                 )
