+++ src/Syntax/Substitution.hs
@@ -9,7 +9,7 @@ import           Data.Map hiding                ( map )
 import           Data.Set                       ( Set )
 import qualified Data.Set as Set
 import           Data.Loc                       ( Loc )
-import           GCL.Common
+import           GCL.Common hiding              ( Substitutable ( subst ) )
 import           GCL.WP.Util                    ( declaredNames )
 import           Syntax.Abstract.Types
 import           Syntax.Common
